requirements.txt
tensorflow-gpu==1.4.1
scikit-learn
pandas
numpy
matplotlib
docopt
